+++ lib/src/pdf_manipulator.dart
@@ -68,6 +68,14 @@ class PdfManipulator {
     return PdfManipulatorPlatform.instance.pdfWatermark(params: params);
   }
 
+  /// Provides pdf file pages size info.
+  ///
+  /// Returns List<PageSizeInfo> for pages size info or null if operation was cancelled.
+  /// Throws exception on error.
+  Future<List<PageSizeInfo>?> pdfPagesSize({PDFPagesSizeParams? params}) {
+    return PdfManipulatorPlatform.instance.pdfPagesSize(params: params);
+  }
+
   /// Cancels running manipulations.
   ///
   /// Returns the cancelling message.